Split script-types to new crate.
