+++ utils/ayaka-bindings-types/Cargo.toml
@@ -5,7 +5,7 @@ authors = ["Berrysoft <Strawberry_Str@hotmail.com>"]
 edition = "2021"
 
 [dependencies]
-ayaka-script = { path = "../ayaka-script" }
+ayaka-script-types = { path = "../ayaka-script-types" }
 fallback = "0.1"
 serde = { version = "1.0", features = ["derive"] }
 log = { version = "0.4", features = ["serde"] }
