+++ utils/ayaka-bindings/src/lib.rs
@@ -4,7 +4,7 @@
 #[doc(no_inline)]
 pub use ayaka_bindings_types::*;
 #[doc(no_inline)]
-pub use ayaka_script::{RawValue, ValueType};
+pub use ayaka_script_types::{RawValue, ValueType};
 #[doc(no_inline)]
 pub use log;
