+++ utils/ayaka-runtime/src/lib.rs
@@ -21,7 +21,9 @@ mod settings;
 #[doc(no_inline)]
 pub use anyhow;
 #[doc(no_inline)]
-pub use ayaka_script::{log, RawValue};
+pub use ayaka_script::log;
+#[doc(no_inline)]
+pub use ayaka_script_types::RawValue;
 pub use config::*;
 pub use context::*;
 #[doc(no_inline)]
