+++ utils/ayaka-script-types/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "ayaka-script-types"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+rt-format = { version = "0.3", optional = true }
+
+[dev-dependencies]
+serde_yaml = "0.9"
+
+[features]
+default = []
