+++ utils/ayaka-script-types/src/lib.rs
@@ -1,0 +1,12 @@
+//! The types used by script and text parsers.
+
+#![warn(missing_docs)]
+#![deny(unsafe_code)]
+
+mod exec;
+mod primitive;
+mod text;
+
+pub use exec::*;
+pub use primitive::*;
+pub use text::*;
