+++ utils/ayaka-script/src/lib.rs
@@ -6,11 +6,9 @@
 #![feature(once_cell)]
 
 mod exec;
-mod primitive;
 mod text;
 
 pub use exec::*;
 #[doc(no_inline)]
 pub use log;
-pub use primitive::*;
 pub use text::*;
