+++ utils/ayaka-script/src/text.rs
@@ -1,6 +1,7 @@
 //! The text parser.
 
-use crate::exec::*;
+use crate::*;
+use ayaka_script_types::*;
 use regex::Regex;
 use std::{error::Error, fmt::Display, iter::Peekable, str::CharIndices, sync::LazyLock};
 
@@ -186,45 +187,6 @@ impl Display for ParseErrorType {
 
 /// The [`std::result::Result`] when parsing [`Text`].
 pub type ParseResult<T> = std::result::Result<T, ParseError>;
-
-/// A part of a line, either some texts or a command.
-#[derive(Debug, PartialEq, Eq)]
-pub enum Line {
-    /// Raw texts.
-    Str(String),
-    /// A command. See [`Command`].
-    Cmd(Command),
-}
-
-/// A TeX-like command in the text.
-#[derive(Debug, PartialEq, Eq)]
-pub enum Command {
-    /// `\ch{}{}`
-    ///
-    /// Controls the current character.
-    Character(String, String),
-    /// `\exec{}`
-    ///
-    /// Executes a program and calculates the return value into text.
-    Exec(Program),
-    /// `\switch{}{}{}`
-    ///
-    /// A switch.
-    Switch {
-        /// The text of the switch.
-        text: String,
-        /// The action after choosing the switch,
-        action: Program,
-        /// The expression determines whether the switch is enabled.
-        enabled: Option<Program>,
-    },
-    /// Other custom commands.
-    Other(String, Vec<String>),
-}
-
-/// A collection of [`Line`].
-#[derive(Debug, Default, PartialEq, Eq)]
-pub struct Text(pub Vec<Line>);
 
 const fn is_special_char(c: char) -> bool {
     matches!(c, '\\' | '{' | '}' | '/')
@@ -683,7 +645,7 @@ mod test_rich_lexer {
 
 #[cfg(test)]
 mod test_parser {
-    use crate::{text::*, *};
+    use crate::text::*;
 
     #[test]
     fn basic() {
